ob-set-defaults: viddle-server can come in flavors, too
Also fix PREFIX problem exposed in older test case data.

+++ ob-set-defaults.bats
@@ -227,8 +227,24 @@ fi
     exit 1
   fi
 
-  # FIXME: following tests are ugly because they use a future version
-  # of cef, so must explicitly specify cef.
+  # Add --asan
+  rm -rf debian
+  cp -a gs4.4-mz3.30 debian
+  ob-set-defaults --g-speak 4.7 --asan
+  if ! diff -ur gs4.7-mz3.30-as debian
+  then
+    echo "ob-set-defaults --g-speak 4.7 --asan did not give expected result on mezzanine"
+    exit 1
+  fi
+
+  # Undo adding --asan
+  ob-set-defaults --g-speak 4.4 --cef 3112
+  if ! diff -ur gs4.4-mz3.30 debian
+  then
+    echo "ob-set-defaults --g-speak 4.4 --cef 3112 did not give expected result on mezzanine"
+    exit 1
+  fi
+
   rm -rf debian
   cp -a debian-mezz322-gs330 debian
   ob-set-defaults --g-speak 3.99 --mezz 5.88 --cef 3112
